Shift function that generates brands.csv file into its own module
- update parser_kfc.js comments and db.sql
